Added useCanvasExport composable
